some changes within entities

+++ src/Contact-zoo-at-home.Core/Entities/Pets/Pet.cs
@@ -102,7 +102,7 @@ namespace Contact_zoo_at_home.Core.Entities.Pets
 
         #region Rating
 
-        public decimal Rating { get; private set; }
+        public float Rating { get; private set; }
 
         public int RatedBy { get; private set; }
 
@@ -111,7 +111,7 @@ namespace Contact_zoo_at_home.Core.Entities.Pets
         /// </summary>
         /// <param name="addingRateMark">Mark to add to the current rating.</param>
         /// <returns>returns new rating.</returns>
-        public decimal AddToRating(decimal addingRateMark) // ToDo: Check calculations
+        public float AddToRating(float addingRateMark) // ToDo: Check calculations
         {
             if (RatedBy == 0)
             {
@@ -119,7 +119,7 @@ namespace Contact_zoo_at_home.Core.Entities.Pets
             }
             else
             {
-                this.Rating = (this.Rating + addingRateMark / RatedBy) * ((decimal)RatedBy / RatedBy + 1);
+                this.Rating = (this.Rating + addingRateMark / RatedBy) * ((float)RatedBy / RatedBy + 1);
             }
 
             RatedBy++;
